[debugger]: added temporary keyword to ocl-break-workitem
In the scenario where we have lots of work-items and a
breakpoint is set, e.g.
(gdb) ocl-break-workitem foo.cl:4
(gdb) r
* Breakpoint Hit (1)
...
(gdb) continue

Then the time from the continue to completion can take
lots of time. This is because we break for all work-items
(CPU threads) and evaluate the stopping condition.
This evaluation takes a lot of time (as opposed to simply
running the kernel to completion without a debugger attached).

Therefore, the temporary keyword usage is like the following:
(gdb) ocl-break-workitem foo.cl:4 0 0 0 temporary

It will set a temporary breakpoint for the requested line
and will stop just once. The continue that follows will end
the kernel promptly.

The default stays a permanent breakpoint.

Change-Id: Ie11c5ab8ebbf364d139b96603ea50e1e3882fce1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/175956
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>