[DTrans] [CMPLRLLVM-1164] Part 1 of dope vector analysis for Transpose transformation
This change adds an analyzer for recognizing and collecting info about uses
of dope vector objects. The dope vectors created directly with the candidate
variable are analyzed with checks that the dope vector for the assumed
shape array represents the entire array, and not a subobject.

Change-Id: I39a46b4c2d4cfc62bc465e4d5fd1d9d9191bd6c3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/182757
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Old/split llvm commit ID: 0abf147ffd1b5aa625d033622244ba1ef2f41227