Revert "Make -Werror to not automatically enable ignored warnings. CQ#374963."
This reverts commit 9080f358787d722542ed79365bf32cfb7e9c499e.

From the original test in cq374963 the problem was clang gave an error
for nonliteral format strings where icc and gcc did not.  Unfortunately
clang was right and icc/gcc wrong (and gcc now diagnoses this too).
Somehow the diagnostic framework was changed to prevent the error but
it incorrectly prevents desired errors.

There was obviously some misunderstanding about how -Werror=format=2 is
supposed to work.  That should turn on format-nonliteral and the lit
test was testing that it did not.

Conflicts:
	lib/Basic/Diagnostic.cpp
	lib/Frontend/ASTUnit.cpp

Change-Id: Ia4b64e48f4d87c17a80fad497405adbb85119474
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/183482
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Ammarguellat, Zahira <zahira.ammarguellat@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Old/split clang commit ID: 787dd67a732aac7907161c1d220de677b49ffcbc