[NFC]CMPLRLLVM-656: Split the functions that are shared between linker and compiler
The whole program depends on communication between the linker and the compiler.
LLD can be built standalone, if we are going to pass a function to LLD, then
it won't have a definition and could produce linking issues. This change set
will put the functions shared between the linker and the compiler in another
file to prevent missing symbols and just export into the linker what is needed.

Change-Id: I6a855aebd69cae980bbede9deaf957a641f873a1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/183988
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Old/split llvm commit ID: 5bd52b662cd506a61436bd0f92036dae4020138c