[imf] Don't constant fold calls to functions with imf attributes
This is an experimental patch that looks for the "imf-arch-consistency"
attribute on functions when considering whether they are candidates for
constant folding. If the attribute is present, they are not folded. This
should be extended to check for other imf attributes.

Part of JIRA: CMPLRLLVM-8752

Change-Id: I2b3bf051a151cda3826f1289d3f1ba4c14c43b76
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/186918
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Old/split llvm commit ID: 710e30cef1331c3a052fe023b3a28aa4fc8506a9