[Workaround] Use opencl.compiler.options metadata for -g option
After [19c80dd2] "Pass compile options directly from runtime to
backend" patch, -g flag is not recognized in case of separate
compilation build (when clCompileProgram and clLinkProgram APIs are
used instead of just clBuildProgram). The information about -g flag is
no longer stored anywhere, and when we execute clLinkProgram,
Optimizer forgets about -g and fully optimizes the program.

This patch enables handling of opencl.compiler.options metadata back
after it was nuked by [19c80dd2]. This approach is still incorrect:
the right way of doing this is by handling "optnone" function
attribute, which is generated by Clang for -O0 flag. Several our
passes do not play well with "optnone", so we add this workaround as
an interim solution.

Change-Id: I54f29255e68af8dfa2a229ed3abb77490e7a027b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/185171
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>