[NFC] [DTrans] Migrate SOAToAOS Transformation to use CallBase class instead of CallSite
This step of the migration updates the SOAToAOS functionality. All objects in
these are updated. Some variable names and member functions still use
CS or CallSite to refer to a call to minimize the differences in the changeset.

Also, a few misc updates for other files to finalize the migration for DTrans,
execept for the interface that is used for indirect calls that will handled when
the indirect call pass is migrated.

Change-Id: I099d533a0c91443f573656808ef8a02152ae922b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/187376
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Chupina, Olga <olga.chupina@intel.com>
Reviewed-by: Zhao, Chuck <chuck.zhao@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Old/split llvm commit ID: aecfb21d2c87b9b172b7b8d2e2e2131434d71f43