CMPLRLLVM-8346: Types comparison when selecting targets for indirect calls conversion [XMAIN]
There are cases when the Andersens analysis find the targets for an indirect call
but it doesn't considers some of them due to a type "mismatch". Although the
Andersens analysis can add the fail-safe case, this is not always true. The problem
is if the set is complete (all targets found) and the analysis trim those "mismatch"
then it can reduce the size enough to remove the fail-safe case. This can produce
an incorrect behavior.

This change set add a types comparison during the process of selecting targets. If
the the type of a target is a structure and the fields matches with the indirect
call, then that target will be selected as possible candidate.

Also, the if we see this types "mismatch" during the indirect call conversion,
apply the proper bitcasting to convert the types in the ir.

Change-Id: Id38844ac6702d76e0a0317a9b06002d141474695
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/188241
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Old/split llvm commit ID: fb745a2836b00921d5c8b6f9d4be6d1d15481657