[InstCombine] Set the nuw flag on nsw multiplies when both inputs are known to be non-negative.
If both inputs are non-negative and the nsw flag is set, then the result must also be
non-negative and with no overflow. This is sufficient to allow nuw to be set as well.

I've restricted this to the case where one of the multiply inputs is a constant to save compile
time.

The motivating case is to improve induction variable simplifications's iv-widening when the iv
is signed and multiplied by a constant. The C overflow rules will ensure the nsw flag is set on
the multiply.

In the IndVarSimplify test case the multiply is sign extended. InstCombine is able
to prove the that the induction variable recurrence must always be positive due to the nsw flag
on the add and the recurrence starting at 0. This further allows InstCombine to prove the mul
must be positive since it has an nsw flag and induction variable is known positive. This causes
the sext to be turned into zext. This causes iv-widening to think the induction variable should
be widened as if it were an unsigned value instead of a signed value. If we don't also have the
nuw flag on the mul, this widening will be incomplete. We'll end up with a widened induction
variable that will be truncated to feed the original i32 multiply and the original zext won't
be removed. Negating any benefit of widening the induction variable.

With this patch we'll be able to widen the multiply to the induction variable width and remove
the zext.

Change-Id: I7fd552beff7adb3b922cbf573aedc0f918b9ca8f
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/188699
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Old/split llvm commit ID: 3efc12712bf1effbca51dbbedb335ce01a016429