cmplrllvm-9194: emit a directive for the implicit target-task when target has nowait/depend:
The rules are:
(1) Variables marked with private/firstprivate on "DIR.OMP.TARGET" should also
be in private/firstprivate clause on the implicit "DIR.OMP.TASK".
(2) Variables used in the "QUAL.OMP.MAP" clause on the "DIR.OMP.TARGET" should
be marked as shared on the "DIR.OMP.TASK".
(3) The implicit task should use a new qualifier called "QUAL.OMP.TARGET.TASK"
to distinguish it from a user specified task.
For target constructs with depend clauses, the rules above apply, in addition
to the following:
(4) There should be a "QUAL.OMP.IF" (i32 0) clause associated with the outer
"QUAL.OMP.TASK". Note that this should not be present if both nowait and depend are used on the same target construct in the user code.
(5) The depend clause should be present on the intrinsic for the implicit task,
and not target .
(6) When there is a map-chain in the IR for the the map clause, the computations
for various expressions used in the map-chain are to be emitted after the
implicit task, and before the target directive. This is similar to having an
explicit #pragma omp task in the user code, just before the #pragma omp target.
The difference being that the base pointer of the map-chain needs to be marked
shared on the task (rule 4 above), and any other values used in computation of
the map-chain (like size in the following example) need to be marked
firstprivate on the task.

Change-Id: I90da769038673902eb3b7d18719c26fd6a61f75d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/188722
Reviewed-by: Yu, Jennifer <jennifer.yu@intel.com>
Tested-by: Yu, Jennifer <jennifer.yu@intel.com>
Old/split clang commit ID: e2618149366c7456871783c24209275d741256fa