Allow loads to be in basic block when considering generating early exit.
Currently early exit is not generated if basic block has load
instruction:

  ...
  %4 = load i32, i32* %3, align 4, !tbaa !8
  ...
  %7 = call i64 @_Z12get_local_idj(i32 0) #3
  %8 = icmp ugt i64 %7, 42
  br i1 %8, label %9, label %10

The reason is that the load instruction is considered as an instruction
which has side effect while it's not. This patch allows generating early
exit(including reducing number of iteration in wg loop) even if there is
load instruction in the basic block.

This improves performance of a couple of SYCL benchmarks that
have the pattern mentioned above.

Change-Id: Iad505e316e214a1b34cc90205503675ff8818dc4
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/190245
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Tested-by: Savonichev, Andrew <andrew.savonichev@intel.com>