Fix a crash when the default of a switch is removed
This patch fixes a problem that occurs in LowerSwitch when a switch
statement has a PHI node as its condition, and the PHI node only has two
incoming blocks, and one of those incoming blocks is through an unreachable
default in the switch statement. When this condition occurs, LowerSwitch
holds a pointer to the condition value, but removes the switch block as
a predecessor of the PHI block, causing the PHI node to be replaced.
LowerSwitch then tries to use its stale pointer to the original
condition value, causing a crash.

This change is also being submitted to llorg, but it is needed in xmain
to prevent OCL LIT test failures.

This addresses failures in CORC-4709.

Change-Id: I9cb5b1bce5bccafadd9e203c3536890a9f0d9614
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/190062
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>
Tested-by: Wenju He <wenju.he@intel.com>
Reviewed-by: Tyurin, Evgeniy <evgeniy.tyurin@intel.com>
Reviewed-by: Guggilla, Satish <satish.guggilla@intel.com>
Old/split llvm commit ID: 19e3507290025f9058bf8c7ffe3ac4be8336cd98