[intel_subgroups][OCLVecClone] "Widen" get_sub_group_local_id during OCLVecClone
get_global_id(unsigned)/get_local_id(unsigned) are specially handled by the
OCLVecClone pass already, do the same for the sub_group local id as it has
almost the same semantics.

TODO: Processing of this function is a little bit hacky because it doesn't have
an explicit dimension id, that needs a proper design.

The same change should probably be done for <*>_linear_id family of functions
too.

Change-Id: Idd5d929be5d62ea07dd4d6f1c33cfb0816f5cd7d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/190514
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>