[Intel_VectorVariant] Fix (de-)mangling of variable stride linears arguments
Non-compile-time constant stride is a mangling of the stride value to a linear
argument, not a separate encoding of its own kind.

The change is motivated by decoupling the divergence info from the mangling
encoding in the VectorKind data structure with the aim to re-use it for the
description of widening decisions of OpenCL sub-group functions. While doing
that I've found that the code and the mangling scheme required by the Vector ABI
actually differ and fixed the issue while on it.

Note that it makes no sense in trying to separate the refactoring and the bug
fix because of the tight coupling of the mangling scheme and the data structure design.

Change-Id: Id847a0b5db4d4615940bc95dd438d41df5a9fd18
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/191251
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Old/split llvm commit ID: 53bfa980a27a7641c8975d0c6754db02f3b3843f