Properly account for the frequency of streaming loads and other HLLs.
The LLVM code that gets generated looks roughly like this:
%0:ci64, %1:ci0 = SLD64 %2:i64, %3:i64, 1, %4:ci0
%5:ci64 = COPY %0:ci64
CSA_LIC_WRITE 1, %5:ci64

Previously, we saw that there was a COPY instruction, which is
non-multitriggered, and assigned the frequency of the block to %0, which is
incorrect. This change weakens the condition for COPY so that it assigns the
same frequency to both but doesn't unify it with the block's frequency.

When move elimination happens, the changes means that the actual frequency of
the use (somewhere inside the loop body) will get backpropagated to the original
output of the SLD instruction.

Change-Id: Ifb638959320dc2a3e8557ca811ebbe85ba797d38
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/190916
Reviewed-by: Woodworth, Daniel <daniel.woodworth@intel.com>
Tested-by: Cranmer, Joshua <joshua.cranmer@intel.com>
Old/split llvm commit ID: d10e4603ca2f6575a9345aa0e877ebdea5c0369f