[CMPLRLLVM-8275,CMPLRLLVM-8653] Remove parallel outer loop check for ILPL
This allows the use of __builtin_csa_pipeline_loop inside of loops that
are not explicitly marked with __builtin_csa_parallel_loop or similar.
It does not change the criteria for automatic selection, and it also
doesn't allow the use of __builtin_csa_pipeline_loop on top-level loops
yet because that's a more intrusive code change.

Change-Id: I0d8d4af1ca892f25680d596622771a9d20e4ebe1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/191472
Reviewed-by: Sudarsanam, Arvind <arvind.sudarsanam@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>
Tested-by: Woodworth, Daniel <daniel.woodworth@intel.com>
Old/split llvm commit ID: 995fca7b7ddbd017b01049b48087c15894ac5374