[VPlan][OVLS] Don't try to interpret DA result inside collectMemrefs
Don't try to interpret VPlanDivergenceAnalysis result inside IR-agnostic
VPlanVLSAnalysis::collectMemrefs method. This is the job of createVLSMemref
method or OVLSMemref subclasses to check for specific shape kinds if necessary.

The benefit of this change is that it prevents early bailout in case of
HIR-based VPlan, so that the VLS analysis and G2S optimization have a chance to
succeed even if VPlanDivergenceAnalysis fails to analyse memory access.

Change-Id: I29313b23d98503135328faf8204099a1d66d0694
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/192211
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Tested-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Old/split llvm commit ID: e39148dcc9f77928342bb2b4adbd59dd562065c4