[VPO] Update renaming logic for handling globals in target-maps.
The renaming code is invoked so that the CodeExtractor creates
params for global variables and constant exprs in the outlined
function.

The problem was that when there are multiple globals and
constant-exprs on a directive, the replacement of one global would
cause breaking of all existing constant-exprs into Instructions,
causing the WRegion, which still points to the original
constant-exprs, to be invalid.

Example:

  region.entry...["MAP"... @x, %1, @y, gep(@y, 0) ]

Before this change, renaming of @x would cause the IR to look like:

  %x = ...
  %y.gep = (@y, 0)
  region.entry...["MAP"... %x, %1, @y, %y.gep ]

While the WRegion still points to gep(@y, 0), which doesn't exist in
the IR anymore. The fix is to create all renamed values in one go,
and then at the end, replace all original values with the renamed ones
separately.

Change-Id: I47e15bfdd1fb34e1ff2ccabf1a9d32a50572d750
JIRA: CMPLRLLVM-9447
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/192576
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Tested-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>
Old/split llvm commit ID: 7067d1619c8efa1a84f5756b540371744ed68869