Out of order clWaitForEvents: add master thread to pool
In out of order queue, clWaitForEvents does not add master thread to
execution pool in the case that event command is not ready to execute
and pCmdDesc is not a valid pointer. This patch adds master thread to
execution pool if pCmdDesc type is CL_DEV_CMD_INVALID.
Jira is CORC-4284.

Change-Id: Ic4d94db43b119dae9baf8130b63e7854ad64c782
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/193058
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>