CMPLRLLVM-9183: Prevent devirtualization on downcasted pointers [XMAIN]
This changeset checks if there is a possible downcasting and prevents
the devirtualization process. For example:

  class Base ...

  class DerivedA : public Base ...

  class DerivedB : public Base ...

  void bar(Base *ptr) {
    DerivedB *ptr2 = (DerivedB *) ptr;
    ptr2->foo();
  }

The downcasting of ptr (ptr2) from Base to DerivedB is incorrect since
devirtualization will assume that ptr2->foo() is a call to DerivedB::foo.
This changeset will identify the downcasting and if it is found then
remove the information used to devirtualize the call site. the indirect
call won't be converted into a direct call.

Change-Id: Ibabee7c5771da1d93dbb40d9406924a6fa98cbbe
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/193517
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>
Old/split llvm commit ID: a6dac8ae87e3fb11ec52bd5555fbe6e7db4e1486