CMPLRLLVM-9505: Support for /lib when is included in an at (@) file [LLD]
The equivalent for MS-LIB in lld is "lld-link /lib". MS-LIB supports
response file (@). This means that it loads the command from a temporary
file to do the linking. In lld "/lib" should be the first argument of the
command, but when is wrapped in an at-file the linker finds that this file
is the first argument.

This change set will check if there is a response file, and if so then expand
it and check the first argument in the expanded command. If the first
argument is /lib then it calls the process of generating a library, else
continue on the regular path.

Change-Id: I66ec660c8d7f559c187e17e77b2dbbdac4ac34ef
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/193946
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>
Old/split lld commit ID: e06098e92483e812d8f68d2a3f339aa831086341