[VPlan] Represent and update validity of underlying IR information for VPValues.
While building a HCFG from incoming LLVM-IR or HIR, every VPValue
has an underlying IR value attached to it. This patch adds a
central mechanism to represent the validity of this mapping and
update it in the cases where a VPValue is modified or updated.
Currently invalidation of underlying IR is done only when operands
of a VPUser are replaced. This validity information can be used
at any stage of VPlan pipeline, and is expected to be kept up-to-date.
Non-VPInstruction VPValues can never be invalidated since they represent
entities outside loop being vectorized.

JIRA : CMPLRLLVM-9250

Change-Id: I00ea6c9bcb49970625223d16926db7da2d07d976
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/194179
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>
Old/split llvm commit ID: 792258342ca5a5a8e7365ff4630409ec4eafeb32