Disable shutdown mechanism on Linux.
There is so called shutdown mechanism in OpenCL CPU implementation,
which is responsible for releasing all OpenCL object which user
allocates and doesn't release(release using clRelease* methods) before
exiting from application. In other words all the leaked OpenCL objects
will be released at process shutdown time.

This mechanism is enabled only for Linux for non-FPGA setup. The
functionality it provides is not required by the OpenCL specification
and there is no known customer request for it. But it leads to the
problems in real application quite often due to problems in the
mechanism itself. For example, some applications just hang at exit
instead of finishing with leaks. Some crashes due to bugs. Those
problems are very hard to debug because environment the logic works
in is very specific - multi-thread multi-library process shutdown.

Change-Id: I487bfff55a276a491d9e1f4d26ef2d38376cfa16
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/194268
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>
Tested-by: Romanov, Vlad <vlad.romanov@intel.com>