Add -mintrinsic-promotion cc1 option to permit feature promotion.
Add a cc1 option to allow a function that calls an intrinsic to have its
feature list promoted. This works in the semantic analysis step, however
the error is left in the CodeGen part to minimize the amount of
conflicts with community.

Change-Id: Ic2279615acc1c4168d6214a6e8eca9779b33c85b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/195029
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Old/split clang commit ID: d12cd1f416a87820fb03770f355cde924640b419