[VPlan][VLS][2/5] Get rid of haveSameNumElements method in client interface
The method doesn't really depend on the type of client, and there's no real need
for the method to be virtual.

The functionality is trivial, so it is clearer to check for equality in-place
instead of having an opaque method for that.

Change-Id: I9918ee63f9fa95fb67a4dbc1cc0ba5ecb8402b46
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/195127
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Tested-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>
Old/split llvm commit ID: 7ddd3cb65df942ebde3365fe03fd0842b4c2a0c1