CMPLRLLVM-9556: Support for reading archive from thin archives made with GNU (Part 1)
The offset of archives inside thin archives made with GNU are presented
as /XX:YY. The XX is the offset of the file name in the string table.
For example:

 /arch1.a
 /0:84    .....

The thin archive presented before has "arch1.a" in the string table
starting at offset 0. The number after the colon can be ignored since
it is used by the member.

This change set basically removes everything after the colon to collect
the offset the member.

Change-Id: Iafa6c9804041e4ee69d05db6d3c6d88ab6f9a2af
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/195196
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>
Old/split llvm commit ID: d4da13df6f5b0736bb847142cdc36fd09bde6bfc