[X86] Fix handling masked load with avx512 and prefer-vector-width=256.
This line was recently changed in community from Legal to Custom to simplify AVX2 masked load handling.
But if we have VLX we don't want to use the AVX2 gather so should go back to Legal. This usually worked
because useAVX512Regs() block later in the file sets it to Legal for VLX. But that block doesn't
execute with -mprefer-vector-width=256.

I'll be pushing the same fix to community, but I think it does bad things to codegen that should be fixed.

Change-Id: I40b8c4e30f158d79430fdbe5da93650e453f4477
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/199253
Reviewed-by: Kaylor, Andrew <andrew.kaylor@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>
Old/split llvm commit ID: 6d1b546621f39470a76f650e57c84165b8025080