[FPGA][EyeQ] Force FPGA/EyeQ emulating modes in compiler config
With SPIR-V to LLVM IR transformation target triple is set as
spir-unknown-unknown, which is making backend thinking that
LLVM IR is not consumed for FPGA device (even if it really is).

Now we are not looking to the triple to decide for what device
we are going to optimize LLVM IR, instead this information is
being obtained from runtime.

Change-Id: I68d4cd8ab30ca9f85b9f6bf6ee668a22d2871c28
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/199389
Reviewed-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>
Tested-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>