Enable support for SYCL FPGA Pipes
SYCL Pipes are enabled as SPIR-V Program Scope pipes, but due to
various design problems, LLVM IR that we get from SPIR-V translator is
not flexible enough to make a robust implementation.

With the current design, program scope pipe is emitted as a global
struct (pipe storage) that consists of three i32 values: size,
alignment and capacity.

Following the design for OpenCL FPGA channels, SYCLPipesHack pass
creates new global variable of type %opencl.pipe_rw_t, initilizes it
with a backing store, and then replaces the original global struct
with this new opaque type.

This design is incomplete and fragile: if an IR heavily optimized, we
may not be able to replace one type with another properly. Proper
solution would be possible if we change IR representation of program
pipes, and add a payload to the storage struct.

Change-Id: I4a19eaf31990cebb26f124e5cb10c3214729df1f
Signed-off-by: Andrew Savonichev <andrew.savonichev@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/199429
Reviewed-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>