[VPlan][DPC++] Fix the compfail in DPC++ rodinia benchmark
JIRA:CMPLRLLVM-9893

This change adds the mapping for the masked float version of SVML sqrt()
function.
One of the calls to 'sqrt' function in the rodinia-cfd benchmark is
under the 'if' branch and hence we require a masked version of SVML sqrt
call. This change adds the entry for that call in the
Intel_SVMLEmitter.cpp file.

TODO: Modify the TLI->isFunctionVectorizable() interface to also check
if the masked version of the call can be vectorized. This would prevent
 us from generating junk function calls and safely fallback to
serial-call execution.

Change-Id: If4236e7cf06abca53513847ff03d64cb90004f74
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/201130
Reviewed-by: Saito, Hideki <hideki.saito@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>
Old/split llvm commit ID: 3fa85bee67e7df9fabfb8491db6e3680580fbcbe