[VPlan][LoopCFU] Properly handle non-phi live-outs of the inner loop
Previous handling (using the mask for the next iteration) only worked for the
recurrence-phi live-outs, but not for the general case. Changed the
implementation to always introduce a new phi node specifically for the live-out
value that works with any def.

Note, that that introduces an additional phi-node for the recurrence-phi
live-out but it doesn't look that bad and the cost of supporting this as a
special case is too high.

Change-Id: Ia8f74ad861bd493ca8049e21751363d3c2c01025
Jira: CMPLRLLVM-9643
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/201907
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>