Recognize pcmpgtw from (mul (and (shr x, 15), 65537), 65535) on 32-bit elements.
This occurs as part of an absolute value of two 16-bit values
packed into a single 32-bit integer that has been vectorized.
The whole sequence is:

y = ((x >> 15) & 0x10001) * 65535;
x -= y;
x ^= y;

Ideally we could use pabsw, but if the low 16-bits of x are
0x8000 then the subtract in the original code will borrow from
the upper 16-bits. This must not occur in the original program,
but we lack the facilities to prove it.

Recognizing the pcmpgtw at least saves us the costly 32-bit
multiply (two serialized 5 cycle uops) and reduces the
instruction count.

Change-Id: I4c6b981d661704fd9b3ce3ecd3229d0c2cfc9f42
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/202100
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>