[CMPLRLLVM-9947] Add pass to enable more fastcc calls for -m32 + PGO builds
This adds a small pass that makes address taken functions become a wrapper
function that forwards its arguments to a function that can have its
calling convention optimized for cases where there are direct calls that
could benefit from a faster calling convention. This is being limited to
PGO cases because the indirect calls would have been specialized in this
situtaion, so most calls would be a direct call.

Change-Id: I8e3ef85c0b3abadf6d4aaaae0344e45a7b614d47
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/202111
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>