[VPlan] Teach VPOVectorizationLegality about uniform loads identified by VPlan DA.
For outer loop vectorization, Legality is unable to correctly identify
uniformity and unit-stridedness for inner loop instructions. This patch
transfers the knowledge about uniform loads from VPlan DA to Legality
during codegen which enables generation of scalar loads instead of gathers
for uniform pointers in vector code. GEPs that are incorrectly marked by
Legality as unit-stride pointers for outer-loop vectorization are also
corrected.

This is a temporary knowledge transfer until we completely switch to
VPValue-based codegen in vectorizer.

Change-Id: I4e74df9498e3b819443ac597acbf76bf05107e1b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/202286
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>