[VPlan DA] Don't crash propagating condition divergence if exit block processed before latch
If latch of a while-loop has single successor (header) then RPOT can process
loop's exit block before the latch is processed, resulting in missing defintion
for the header for which we asserted before this change. Start to mark the loop
as divergent instead of the assert.

Patch by Matt Masten, I've just cleaned up the test a little bit.

Change-Id: Ibf0df8d7342d57d3fa2cdf9635e149dde2562d56
Jira: CMPLRLLVM-9257
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/202305
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>