[VPlan] Fix serialized codegen for masked uniform loads.
For loads from uniform pointers VPlan serializes the instruction
and generates scalar loads. However masking should also be accounted
for before serializing. This patch fixes codegen to check for
mask when a uniform load is serialized. We generate a single scalar
load to the uniform pointer predicated by a not all-zero check on
the mask value.

JIRA : CMPLRLLVM-9985

Change-Id: I7d1c848eb4902f554ae97a8e7b8b7359b588ef37
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/202839
Reviewed-by: Pushkar Ratnalikar <pushkar.v.ratnalikar@intel.com>
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>