[VPlan] - Fix lookup of function to determine if it is  vectorizable in presence of a mask.
Check mask value in addition to the function-name, when we want to
determine if the function is vectorizable. If the masked version of the
function is not available, we want fallback and serialize the execution
of that function call.

Changes migrated to mono-repo.
Previous CR- https://git-amr-2.devtools.intel.com/gerrit/#/c/201558/

Jira: CMPLRLLVM-9929
Change-Id: Iadc13721a7fd4572ad2ace012b0c8782e6fd1ef6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/202910
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>