Resolve build failure after da60fc813c2 Changing representation of .cv_def_range directives in Codeview debug info assembly format for better readability.
This is just a temporary hack, just to work around a flaw (duplicated defnition of symbols) in our own code to get STIDebug.cpp to compile. We should, however, fix this up but given the gap we don't have much time now. So first we need this patch to unblock the pulldown, then we see if we can fix it, and this patch will be reverted afterwards.

Change-Id: Ia1e466db4980ee1e3e12b15cb9e24d251ce781fe
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/203214
Reviewed-by: Tianqing Wang <tianqing.wang@intel.com>
Tested-by: Tianqing Wang <tianqing.wang@intel.com>