[VPlan][NFC] Use a separate flag in VPValue to identify validity of underlying llvm::Value.
The current setup for invalidating a VPValue drops its corresponding
pointer to the underlying llvm::Value. This patch modifies this by
adding a new boolean flag in VPValue which indicates if the underlying
Value is invalidated. The original underlying Value maybe needed for
later use-cases in VPlan. Current IR-based vector codegen is also
updated to check for validity of VPValue instead of directly peeking
at Value pointer.

Change-Id: I437453d588981446486ab3cee509b2f7b753771a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/203351
Tested-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Reviewed-by: Prokofiev, Pavel G <pavel.g.prokofiev@intel.com>