Change a way how OpenCL libraries are loaded
Summary:
Some of OpenCL libraries (for example libOclCpuBackEnd.so) are loaded
on linux via dlopen() function. It searches for a library by name
in a path specified in LD_LIBRARY_PATH. First library found is being
loaded.

It could happen that some of OpenCL libraries weren't be loaded.
For example, if specified ICD on user's machine was pointing to
symlink of OpenCL RT entry point and actual library placement wasn't
presented in LD_LIBRARY_PATH. In that case OCLBackend library was failed
to be loaded.

With this patch actual string passed in dlopen() includes not only
a library name, but also a full path to it. It's being calculated as
a relative path to a library from where the function was called. We
expect, that a callee library is having the same path as the caller.

Change-Id: Ieb3d203f0c7242e0434385871134d9eaa16d6b5a
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/205017
Tested-by: Sidorov, Dmitry <dmitry.sidorov@intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>
Reviewed-by: Savonichev, Andrew <andrew.savonichev@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>