Implement an OCL specific loop idiom recognize
This pass implements a loop idiom recognizer that transforms simple loop
into a non-loop form, e.g. memcpy. In cases that this kicks in,
it can be a significant performance win.

The standard LLVM idiom pass couldn't recognize some OCL loops, because it
have some limitations, e.g. for a block to be unconditionally executed,
it has to dominate all the exit blocks of the loop. The code change is big
to modify standard idiom to recognize OCL loop, the risk is high.

This pass will generate dead loops, so it should combine a loop-deletion
pass to use, it does a lot of simplifications to simplify pattern match work.

Change-Id: Iaaea0ac723af60f355f8c2525ee10d87985ac864
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/203678
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Tianyang Zhu <tianyang.zhu@intel.com>
Reviewed-by: Dai, Chunyang <chunyang.dai@intel.com>